// Ticket_Concert FE Admin/interfaces/NguoiDung.ts
export interface NguoiDung {
    IDNguoiDung: string;
    HinhAnh?: string | null;
    TenNguoiDung: string;
    Email: string;
    SoDienThoai: string;
    GioiTinh?: "Nam" | "Nữ" | "Khác" | null;
    NgaySinh?: string | null;
    MatKhau: string;
    QuyenHan: "Admin" | "User";
    TrangThai : "Hoạt động" | "Khoá"
  }
  